+++ src/main/java/fi/vm/sade/valinta/kooste/viestintapalvelu/route/impl/HyvaksymiskirjeetServiceImpl.java
@@ -22,6 +22,7 @@ import fi.vm.sade.valinta.kooste.external.resource.valintatulosservice.ValintaTu
 import fi.vm.sade.valinta.kooste.external.resource.viestintapalvelu.ViestintapalveluAsyncResource;
 import fi.vm.sade.valinta.kooste.parametrit.ParametritParser;
 import fi.vm.sade.valinta.kooste.parametrit.service.HakuParametritService;
+import fi.vm.sade.valinta.kooste.sijoitteluntulos.dto.SijoittelunTulosProsessi;
 import fi.vm.sade.valinta.kooste.util.HakemusWrapper;
 import fi.vm.sade.valinta.kooste.valvomo.dto.Poikkeus;
 import fi.vm.sade.valinta.kooste.viestintapalvelu.Hakijapalvelu;
@@ -43,6 +44,7 @@ import io.reactivex.schedulers.Schedulers;
 import io.reactivex.subjects.PublishSubject;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang3.time.DurationFormatUtils;
+import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -54,6 +56,7 @@ import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -279,6 +282,83 @@ public class HyvaksymiskirjeetServiceImpl implements HyvaksymiskirjeetService {
                 .subscribe(
                         letterBatch -> letterBatchToViestintapalvelu().accept(new Object[] { letterBatch, prosessi, hyvaksymiskirjeDTO }),
                         throwable -> logErrorAndKeskeyta(prosessi, throwable, hakuOid, hakukohdeOid));
+    }
+
+    @Override
+    public void hyvaksymiskirjeetHaulle(String hakuOid, String asiointikieli, SijoittelunTulosProsessi prosessi, String defaultValue) {
+        LOG.info("Aloitetaan haun {} hyväksymiskirjeiden luonti asiointikielelle {} hakemalla hyväksytyt koko haulle", hakuOid, prosessi.getAsiointikieli());
+        prosessi.setKokonaistyo(1);
+        valintaTulosServiceAsyncResource.getKoulutuspaikalliset(hakuOid)
+                .flatMap(valintatulokset -> tarjontaAsyncResource.haeHaku(hakuOid)
+                        .flatMap(haku -> {
+                            List<String> hakemusOids = valintatulokset.getResults().stream().map(HakijaDTO::getHakemusOid).collect(Collectors.toList());
+                            return StringUtils.isEmpty(haku.getAtaruLomakeAvain())
+                                    ? applicationAsyncResource.getApplicationsByhakemusOidsInParts(hakuOid, hakemusOids, ApplicationAsyncResource.DEFAULT_KEYS)
+                                    : ataruAsyncResource.getApplicationsByOids(hakemusOids);
+                        })
+                        .flatMap(hakemukset -> {
+                            List<HakemusWrapper> asiointikielisetHakemukset = hakemukset.stream()
+                                    .filter(h -> asiointikieli.equalsIgnoreCase(h.getAsiointikieli()))
+                                    .collect(Collectors.toList());
+                            Set<String> asiointikielisetHakemusOids = asiointikielisetHakemukset.stream().map(HakemusWrapper::getOid).collect(Collectors.toSet());
+                            List<HakijaDTO> asiointikielisetValintatulokset = valintatulokset.getResults().stream()
+                                    .filter(v -> asiointikielisetHakemusOids.contains(v.getHakemusOid()))
+                                    .collect(Collectors.toList());
+
+                            Map<String, MetaHakukohde> hakukohteet = hyvaksymiskirjeetKomponentti.haeKiinnostavatHakukohteet(asiointikielisetValintatulokset);
+                            ParametritParser haunParametrit = hakuParametritService.getParametritForHaku(hakuOid);
+
+                            return Observable.fromIterable(hakukohteet.values().stream().map(MetaHakukohde::getTarjoajaOid).distinct()::iterator)
+                                    .flatMap(tarjoajaOid -> organisaatioAsyncResource.haeHakutoimisto(tarjoajaOid)
+                                                    .map(toimisto -> Pair.of(tarjoajaOid, toimisto.flatMap(t -> Hakijapalvelu.osoite(t, asiointikieli)))),
+                                            1)
+                                    .<Map<String, Optional<Osoite>>>collect(HashMap::new, (map, pair) -> map.put(pair.getLeft(), pair.getRight()))
+                                    .toObservable()
+                                    .map(hakijapalveluidenOsoite -> HyvaksymiskirjeetKomponentti.teeHyvaksymiskirjeet(
+                                            koodistoCachedAsyncResource::haeKoodisto,
+                                            hakijapalveluidenOsoite,
+                                            hakukohteet,
+                                            asiointikielisetValintatulokset,
+                                            asiointikielisetHakemukset,
+                                            null,
+                                            hakuOid,
+                                            Optional.of(asiointikieli),
+                                            defaultValue,
+                                            hakuOid,
+                                            "hyvaksymiskirje",
+                                            parsePalautusPvm(null, haunParametrit),
+                                            parsePalautusAika(null, haunParametrit),
+                                            true,
+                                            true
+                                    ));
+                        }))
+                .flatMap(viestintapalveluAsyncResource::viePdfJaOdotaReferenssiObservable)
+                .flatMap(letterResponse -> Observable.interval(1, TimeUnit.SECONDS)
+                        .flatMap(i -> viestintapalveluAsyncResource.haeStatusObservable(letterResponse.getBatchId())
+                                .flatMap(letterBatchStatus -> {
+                                    if ("error".equals(letterBatchStatus.getStatus())) {
+                                        return Observable.error(new RuntimeException("Viestintäpalvelun statuspyyntö palautti virheen"));
+                                    }
+                                    if ("ready".equals(letterBatchStatus.getStatus())) {
+                                        return Observable.just(letterResponse.getBatchId());
+                                    }
+                                    return Observable.empty();
+                                }))
+                        .take(1))
+                .timeout(780, TimeUnit.MINUTES, Observable.just("timeout"))
+                .subscribe(
+                        batchId -> {
+                            // TODO timeout handling
+                            prosessi.setDokumenttiId(batchId);
+                            prosessi.inkrementoi();
+                            LOG.info("Haun hyväksymiskirjeet valmiit");
+                        },
+                        error -> {
+                            LOG.error("Haun hyväksymiskirjeiden muodostaminen ei onnistunut", error);
+                            prosessi.inkrementoiOhitettujaToita();
+                            prosessi.getPoikkeukset().add(Poikkeus.koostepalvelupoikkeus("Hyväksymiskirjeiden muodostaminen ei onnistunut.\n" + error.getMessage()));
+                        }
+                );
     }
 
     private List<HakijaDTO> paatteleTarpeellisetKirjeet(List<HakijaDTO> hakijat, List<HakemusWrapper> hakemukset, HyvaksymiskirjeDTO hyvaksymiskirjeDTO) {
